# Remove usage of bytes.Buffer and bytes.Reader